+++ CHANGELOG.md
@@ -12,6 +12,7 @@ _general description_
 - **added** - `<concepts>` and `<ranges>` check and import
 - **added** - simple logging system with different levels and color support
 - **added** - unit test framework without macros
+- **added** - time util
 
 ---